test: move refresh close test to connection
Exercise the connection-owned authorizer with a gated refresh credential.
Expose only test-gated helpers for deterministic refresh timing and setup.
